requirements.txt
bitsandbytes==0.42.0
transformers==4.36.2
peft==0.7.1
accelerate==0.26.1
wandb==0.16.2
datasets==2.16.1
trl==0.7.10
deepmerge==1.1.1
flask==3.0.1
waitress==2.1.2
python-dotenv==1.0.1
pandas==2.0.3
jinja2==3.1.3
html2text==2020.1.16
